+++ website/index.html
@@ -4,7 +4,58 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>QwenPaw — Works for you, grows with you.</title>
+    <meta
+      name="description"
+      content="QwenPaw 是运行在你自己环境中的 AI 个人助理：本地或云端部署，支持钉钉、飞书、微信、Discord、Telegram 等多端接入，通过 Skills 自由扩展能力。Your personal AI assistant that runs in your own environment."
+    />
+    <link rel="canonical" href="https://qwenpaw.agentscope.io/" />
     <link rel="icon" href="/qwenpaw-symbol.svg" type="image/svg+xml" />
+
+    <meta name="google-site-verification" content="kRy3rs7ns4DbvoAKOafWZldfesOJOut0I82DwT7ombg" />
+    <meta name="msvalidate.01" content="9F648E7030DEAA7958D1CA55FDEB04E1" />
+    <meta name="baidu-site-verification" content="codeva-q5MuCpIdd2" />
+    <meta property="og:type" content="website" />
+    <meta property="og:site_name" content="QwenPaw" />
+    <meta
+      property="og:title"
+      content="QwenPaw — Works for you, grows with you."
+    />
+    <meta
+      property="og:description"
+      content="Your personal AI assistant that runs in your own environment. Local or cloud deployment, multi-channel access (DingTalk, Feishu, WeChat, Discord, Telegram), extensible via Skills."
+    />
+    <meta property="og:url" content="https://qwenpaw.agentscope.io/" />
+    <meta
+      property="og:image"
+      content="https://qwenpaw.agentscope.io/qwenpaw_ip.png"
+    />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta
+      name="twitter:title"
+      content="QwenPaw — Works for you, grows with you."
+    />
+    <meta
+      name="twitter:description"
+      content="Your personal AI assistant that runs in your own environment. Local or cloud deployment, multi-channel access, extensible via Skills."
+    />
+    <meta
+      name="twitter:image"
+      content="https://qwenpaw.agentscope.io/qwenpaw_ip.png"
+    />
+    <link
+      rel="preload"
+      href="/font/Inter.woff2"
+      as="font"
+      type="font/woff2"
+      crossorigin
+    />
+    <link
+      rel="preload"
+      href="/font/Newsreader.woff2"
+      as="font"
+      type="font/woff2"
+      crossorigin
+    />
   </head>
   <body>
     <div id="root"></div>
